+++ pwm.h
@@ -55,7 +55,7 @@
 
 typedef struct
 {
-    uint32_t ctl;
+    uint64_t ctl;
 #define RPI_PWM_CTL_MSEN2                        (1 << 15)
 #define RPI_PWM_CTL_USEF2                        (1 << 13)
 #define RPI_PWM_CTL_POLA2                        (1 << 12)
@@ -71,7 +71,7 @@ typedef struct
 #define RPI_PWM_CTL_RPTL1                        (1 << 2)
 #define RPI_PWM_CTL_MODE1                        (1 << 1)
 #define RPI_PWM_CTL_PWEN1                        (1 << 0)
-    uint32_t sta;
+    uint64_t sta;
 #define RPI_PWM_STA_STA4                         (1 << 12)
 #define RPI_PWM_STA_STA3                         (1 << 11)
 #define RPI_PWM_STA_STA2                         (1 << 10)
@@ -85,17 +85,17 @@ typedef struct
 #define RPI_PWM_STA_WERR1                        (1 << 2)
 #define RPI_PWM_STA_EMPT1                        (1 << 1)
 #define RPI_PWM_STA_FULL1                        (1 << 0)
-    uint32_t dmac;
+    uint64_t dmac;
 #define RPI_PWM_DMAC_ENAB                        (1 << 31)
 #define RPI_PWM_DMAC_PANIC(val)                  ((val & 0xff) << 8)
 #define RPI_PWM_DMAC_DREQ(val)                   ((val & 0xff) << 0)
-    uint32_t resvd_0x0c;
-    uint32_t rng1;
-    uint32_t dat1;
-    uint32_t fif1;
-    uint32_t resvd_0x1c;
-    uint32_t rng2;
-    uint32_t dat2;
+    uint64_t resvd_0x0c;
+    uint64_t rng1;
+    uint64_t dat1;
+    uint64_t fif1;
+    uint64_t resvd_0x1c;
+    uint64_t rng2;
+    uint64_t dat2;
 } __attribute__((packed, aligned(4))) pwm_t;
