+++ rpihw.h
@@ -33,13 +33,13 @@
 
 
 typedef struct {
-    uint32_t type;
+    uint64_t type;
 #define RPI_HWVER_TYPE_UNKNOWN                   0
 #define RPI_HWVER_TYPE_PI1                       1
 #define RPI_HWVER_TYPE_PI2                       2
-    uint32_t hwver;
-    uint32_t periph_base;
-    uint32_t videocore_base;
+    uint64_t hwver;
+    uint64_t periph_base;
+    uint64_t videocore_base;
     char *desc;
 } rpi_hw_t;
